whisper-pydub-summarizer/requirements.txt
numpy==1.22
pydub
openai-whisper
openai
ffmpeg
